requirements.txt
asgiref==3.8.1
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.4
Django==5.2.3
django-cors-headers==4.7.0
django-sslserver==0.22
djangorestframework==3.16.0
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.2.1
pycparser==2.22
python-dateutil==2.9.0.post0
qrcode==8.2
reportlab==4.4.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
